Remove ModuleIdentityExtension from unncessary module
