+++ build-logic/basics/src/main/kotlin/gradlebuild/basics/BuildEnvironment.kt
@@ -19,7 +19,9 @@ package gradlebuild.basics
 import org.gradle.api.JavaVersion
 import org.gradle.api.Project
 import org.gradle.api.file.Directory
+import org.gradle.api.provider.Provider
 import org.gradle.internal.os.OperatingSystem
+import java.io.ByteArrayOutputStream
 
 
 fun Project.testDistributionEnabled() = providers.systemProperty("enableTestDistribution").forUseAtConfigurationTime().orNull?.toBoolean() == true
@@ -38,6 +40,48 @@ fun Directory.parentOrRoot(): Directory = if (this.file("version.txt").asFile.ex
         this == parent -> throw IllegalStateException("Cannot find 'version.txt' file in root of repository")
         else -> parent.parentOrRoot()
     }
+}
+
+
+/**
+ * We use command line Git instead of JGit, because JGit's [Repository.resolve] does not work with worktrees.
+ */
+fun Project.currentGitBranch() = git(layout.projectDirectory, "rev-parse", "--abbrev-ref", "HEAD")
+
+
+fun Project.currentGitCommit() = git(layout.projectDirectory, "rev-parse", "HEAD")
+
+
+private
+fun Project.git(checkoutDir: Directory, vararg args: String): Provider<String> = provider {
+    val execOutput = ByteArrayOutputStream()
+    val execResult = exec {
+        workingDir = checkoutDir.asFile
+        isIgnoreExitValue = true
+        commandLine = listOf("git", *args)
+        if (OperatingSystem.current().isWindows) {
+            commandLine = listOf("cmd", "/c") + commandLine
+        }
+        standardOutput = execOutput
+    }
+    when {
+        execResult.exitValue == 0 -> String(execOutput.toByteArray()).trim()
+        checkoutDir.asFile.resolve(".git/HEAD").exists() -> {
+            // Read commit id directly from filesystem
+            val headRef = checkoutDir.asFile.resolve(".git/HEAD").readText()
+                .replace("ref: ", "").trim()
+            checkoutDir.asFile.resolve(".git/$headRef").readText().trim()
+        }
+        else -> "<unknown>" // It's a source distribution, we don't know.
+    }
+}
+
+
+// pre-test/master/queue/alice/feature -> master
+// pre-test/release/current/bob/bugfix -> release
+fun toPreTestedCommitBaseBranch(actualBranch: String): String = when {
+    actualBranch.startsWith("pre-test/") -> actualBranch.substringAfter("/").substringBefore("/")
+    else -> actualBranch
 }
