+++ build-logic/build-update-utils/build.gradle.kts
@@ -5,5 +5,5 @@ plugins {
 
 dependencies {
     implementation("com.google.code.gson:gson")
-    implementation(project(":module-identity"))
+    implementation(project(":basics"))
 }
