+++ build-logic/build-update-utils/src/main/kotlin/gradlebuild/buildutils/tasks/UpdateBranchStatus.kt
@@ -22,7 +22,10 @@ import org.gradle.api.Project
 import org.gradle.kotlin.dsl.*
 import java.io.ByteArrayOutputStream
 import java.io.File
-import gradlebuild.identity.extension.ModuleIdentityExtension
+import org.gradle.api.provider.Property
+import org.gradle.api.provider.Provider
+import gradlebuild.basics.toPreTestedCommitBaseBranch
+import org.gradle.api.tasks.Internal
 
 
 /**
@@ -30,10 +33,15 @@ import gradlebuild.identity.extension.ModuleIdentityExtension
  * green-master/green-release to remote repository so that developers can checkout new branches from these tags.
  */
 abstract class UpdateBranchStatus : DefaultTask() {
+    @get: Internal
+    abstract val gradleBuildBranch: Property<String>
+    private
+    val logicalBranch: Provider<String>
+        get() = gradleBuildBranch.map(::toPreTestedCommitBaseBranch)
 
     @TaskAction
     fun publishBranchStatus() {
-        when (project.the<ModuleIdentityExtension>().logicalBranch.get()) {
+        when (logicalBranch.get()) {
             "master" -> publishBranchStatus("master")
             "release" -> publishBranchStatus("release")
         }
@@ -41,7 +49,7 @@ abstract class UpdateBranchStatus : DefaultTask() {
 
     private
     fun publishBranchStatus(branch: String) {
-        val actualBranch = project.the<ModuleIdentityExtension>().gradleBuildBranch.get()
+        val actualBranch = gradleBuildBranch.get()
         println("Publishing branch status of $branch from $actualBranch")
         project.execAndGetStdout("git", "push", "https://bot-teamcity:${System.getenv("BOT_TEAMCITY_GITHUB_TOKEN")}@github.com/gradle/gradle.git", "$actualBranch:green-$branch")
     }
