+++ build-logic/module-identity/src/main/kotlin/gradlebuild/identity/extension/ModuleIdentityExtension.kt
@@ -16,6 +16,7 @@
 
 package gradlebuild.identity.extension
 
+import gradlebuild.basics.toPreTestedCommitBaseBranch
 import gradlebuild.identity.tasks.BuildReceipt
 import org.gradle.api.model.ObjectFactory
 import org.gradle.api.provider.Property
@@ -49,7 +50,7 @@ abstract class ModuleIdentityExtension(val tasks: TaskContainer, val objects: Ob
      *
      * For example, for the pre-tested commit branch "pre-test/master/queue/alice/personal-branch" the logical branch is "master".
      */
-    val logicalBranch: Provider<String> = gradleBuildBranch.map(this::toPreTestedCommitBaseBranch)
+    val logicalBranch: Provider<String> = gradleBuildBranch.map(::toPreTestedCommitBaseBranch)
 
     abstract val gradleBuildCommitId: Property<String>
 
@@ -68,13 +69,5 @@ abstract class ModuleIdentityExtension(val tasks: TaskContainer, val objects: Ob
         tasks.named<Jar>("jar").configure {
             from(createBuildReceipt.map { it.receiptFolder })
         }
-    }
-
-    // pre-test/master/queue/alice/feature -> master
-    // pre-test/release/current/bob/bugfix -> release
-    private
-    fun toPreTestedCommitBaseBranch(actualBranch: String): String = when {
-        actualBranch.startsWith("pre-test/") -> actualBranch.substringAfter("/").substringBefore("/")
-        else -> actualBranch
     }
 }
